+++ api/datadogV1/model_idp_response.go
@@ -78,21 +78,10 @@ func (o IdpResponse) MarshalJSON() ([]byte, error) {
 // UnmarshalJSON deserializes the given payload.
 func (o *IdpResponse) UnmarshalJSON(bytes []byte) (err error) {
 	raw := map[string]interface{}{}
-	required := struct {
+	all := struct {
 		Message *string `json:"message"`
 	}{}
-	all := struct {
-		Message string `json:"message"`
-	}{}
-	err = json.Unmarshal(bytes, &required)
-	if err != nil {
-		return err
-	}
-	if required.Message == nil {
-		return fmt.Errorf("required field message missing")
-	}
-	err = json.Unmarshal(bytes, &all)
-	if err != nil {
+	if err = json.Unmarshal(bytes, &all); err != nil {
 		err = json.Unmarshal(bytes, &raw)
 		if err != nil {
 			return err
@@ -100,13 +89,16 @@ func (o *IdpResponse) UnmarshalJSON(bytes []byte) (err error) {
 		o.UnparsedObject = raw
 		return nil
 	}
+	if all.Message == nil {
+		return fmt.Errorf("required field message missing")
+	}
 	additionalProperties := make(map[string]interface{})
 	if err = json.Unmarshal(bytes, &additionalProperties); err == nil {
 		datadog.DeleteKeys(additionalProperties, &[]string{"message"})
 	} else {
 		return err
 	}
-	o.Message = all.Message
+	o.Message = *all.Message
 	if len(additionalProperties) > 0 {
 		o.AdditionalProperties = additionalProperties
 	}
