+++ api/datadogV1/model_logs_pipelines_order.go
@@ -79,21 +79,10 @@ func (o LogsPipelinesOrder) MarshalJSON() ([]byte, error) {
 // UnmarshalJSON deserializes the given payload.
 func (o *LogsPipelinesOrder) UnmarshalJSON(bytes []byte) (err error) {
 	raw := map[string]interface{}{}
-	required := struct {
+	all := struct {
 		PipelineIds *[]string `json:"pipeline_ids"`
 	}{}
-	all := struct {
-		PipelineIds []string `json:"pipeline_ids"`
-	}{}
-	err = json.Unmarshal(bytes, &required)
-	if err != nil {
-		return err
-	}
-	if required.PipelineIds == nil {
-		return fmt.Errorf("required field pipeline_ids missing")
-	}
-	err = json.Unmarshal(bytes, &all)
-	if err != nil {
+	if err = json.Unmarshal(bytes, &all); err != nil {
 		err = json.Unmarshal(bytes, &raw)
 		if err != nil {
 			return err
@@ -101,13 +90,16 @@ func (o *LogsPipelinesOrder) UnmarshalJSON(bytes []byte) (err error) {
 		o.UnparsedObject = raw
 		return nil
 	}
+	if all.PipelineIds == nil {
+		return fmt.Errorf("required field pipeline_ids missing")
+	}
 	additionalProperties := make(map[string]interface{})
 	if err = json.Unmarshal(bytes, &additionalProperties); err == nil {
 		datadog.DeleteKeys(additionalProperties, &[]string{"pipeline_ids"})
 	} else {
 		return err
 	}
-	o.PipelineIds = all.PipelineIds
+	o.PipelineIds = *all.PipelineIds
 	if len(additionalProperties) > 0 {
 		o.AdditionalProperties = additionalProperties
 	}
