+++ api/datadogV1/model_synthetics_trigger_body.go
@@ -78,21 +78,10 @@ func (o SyntheticsTriggerBody) MarshalJSON() ([]byte, error) {
 // UnmarshalJSON deserializes the given payload.
 func (o *SyntheticsTriggerBody) UnmarshalJSON(bytes []byte) (err error) {
 	raw := map[string]interface{}{}
-	required := struct {
+	all := struct {
 		Tests *[]SyntheticsTriggerTest `json:"tests"`
 	}{}
-	all := struct {
-		Tests []SyntheticsTriggerTest `json:"tests"`
-	}{}
-	err = json.Unmarshal(bytes, &required)
-	if err != nil {
-		return err
-	}
-	if required.Tests == nil {
-		return fmt.Errorf("required field tests missing")
-	}
-	err = json.Unmarshal(bytes, &all)
-	if err != nil {
+	if err = json.Unmarshal(bytes, &all); err != nil {
 		err = json.Unmarshal(bytes, &raw)
 		if err != nil {
 			return err
@@ -100,13 +89,16 @@ func (o *SyntheticsTriggerBody) UnmarshalJSON(bytes []byte) (err error) {
 		o.UnparsedObject = raw
 		return nil
 	}
+	if all.Tests == nil {
+		return fmt.Errorf("required field tests missing")
+	}
 	additionalProperties := make(map[string]interface{})
 	if err = json.Unmarshal(bytes, &additionalProperties); err == nil {
 		datadog.DeleteKeys(additionalProperties, &[]string{"tests"})
 	} else {
 		return err
 	}
-	o.Tests = all.Tests
+	o.Tests = *all.Tests
 	if len(additionalProperties) > 0 {
 		o.AdditionalProperties = additionalProperties
 	}
